learn/fmri_mvpa/fmri_mvpa_enhance/requirements.txt
# 增强版MVPA分析流水线依赖包
# Enhanced MVPA Analysis Pipeline Requirements

# 核心科学计算库 (Core Scientific Computing)
numpy>=1.20.0
scipy>=1.7.0
pandas>=1.3.0

# 神经影像处理 (Neuroimaging Processing)
nibabel>=3.2.0
nilearn>=0.8.0

# 机器学习 (Machine Learning)
scikit-learn>=1.0.0

# 可视化 (Visualization)
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
kaleido>=0.2.1  # for plotly static image export

# 并行处理 (Parallel Processing)
joblib>=1.1.0

# 系统监控和进度条 (System Monitoring and Progress)
psutil>=5.8.0
tqdm>=4.62.0

# 数据类和类型提示 (Dataclasses and Type Hints)
dataclasses>=0.6; python_version<"3.7"
typing-extensions>=3.7.4; python_version<"3.8"

# 日志和配置 (Logging and Configuration)
PyYAML>=5.4.0

# 可选依赖 (Optional Dependencies)
# 取消注释以下行来安装可选功能

# PDF报告生成 (PDF Report Generation)
# weasyprint>=54.0

# 更好的进度条和终端输出 (Better Progress Bars and Terminal Output)
# rich>=10.0.0

# 内存分析 (Memory Profiling)
# memory-profiler>=0.60.0

# 更好的错误追踪 (Better Error Tracking)
# sentry-sdk>=1.0.0

# 数据验证 (Data Validation)
# pydantic>=1.8.0

# 更快的JSON处理 (Faster JSON Processing)
# orjson>=3.6.0

# 网络请求 (HTTP Requests)
# requests>=2.25.0

# 统计分析 (Statistical Analysis)
# statsmodels>=0.12.0

# 更多可视化选项 (Additional Visualization Options)
# bokeh>=2.3.0
# altair>=4.1.0

# Jupyter支持 (Jupyter Support)
# jupyter>=1.0.0
# ipywidgets>=7.6.0

# 开发依赖 (Development Dependencies)
# pytest>=6.2.0
# pytest-cov>=2.12.0
# black>=21.0.0
# flake8>=3.9.0
# mypy>=0.812